Authorisation in controllers

+++ app/Http/Controllers/ServerController.php
@@ -7,8 +7,10 @@ use Illuminate\Http\Request;
 
 class ServerController extends Controller
 {
+
     public function index()
     {
+        $this->authorize('viewAny', Server::class);
         $servers = Server::withCount('users')
             ->paginate(config('constants.PAGINATION_LIST_LENGTH'));
 
@@ -18,12 +20,15 @@ class ServerController extends Controller
     // Show the form to create a new servers
     public function create()
     {
+        $this->authorize('create', Server::class);
         return view('servers.create');
     }
 
     // Store a new servers
     public function store(Request $request)
     {
+        $this->authorize('create', Server::class);
+
         $request->validate([
             'name' => 'required|max:250',
             'capacity' => 'required|integer|min:1',
@@ -41,6 +46,9 @@ class ServerController extends Controller
     public function show($id)
     {
         $server = Server::findOrFail($id);
+        // Manually authorize the 'view' action
+        $this->authorize('view', $server);
+
         $users = $server->users()->paginate(config('constants.PAGINATION_LIST_LENGTH'));
 
         return view('servers.show', compact('server', 'users'));
@@ -50,12 +58,14 @@ class ServerController extends Controller
     public function edit($id)
     {
         $server = Server::findOrFail($id);
+        $this->authorize('update', $server);
         return view('servers.edit', compact('server'));
     }
 
     // Update a specific servers
     public function update(Request $request, Server $server)
     {
+        $this->authorize('update', $server);
         // Validate the inputs
         $request->validate([
             'name' => 'required|string|max:255',
@@ -81,6 +91,7 @@ class ServerController extends Controller
     public function destroy($id)
     {
         $server = Server::findOrFail($id);
+        $this->authorize('delete', $server);
         $server->delete();
 
         return redirect()->route('servers.index');
