+++ app/Http/Controllers/UserController.php
@@ -10,11 +10,13 @@ use Illuminate\Support\Facades\Hash;
 
 class UserController extends Controller
 {
+
     /**
      * Display a listing of the users.
      */
     public function index()
     {
+        $this->authorize('viewAny', User::class);
         $users = User::withCount('servers')
             ->paginate(config('constants.PAGINATION_LIST_LENGTH'));
         return view('users.index', compact('users'));
@@ -25,6 +27,7 @@ class UserController extends Controller
      */
     public function create()
     {
+        $this->authorize('create', User::class);
         return view('users.create');
     }
 
@@ -33,13 +36,14 @@ class UserController extends Controller
      */
     public function store(StoreUserRequest $request)
     {
+        $this->authorize('create', User::class);
+
         User::create([
             'name' => $request->name,
             'email' => $request->email,
             'password' => Hash::make($request->password),
             'role' => $request->role,
         ]);
-
         return redirect()->route('users.index')->with('success', 'User created successfully');
     }
 
@@ -49,6 +53,7 @@ class UserController extends Controller
     public function show($id)
     {
         $user = User::findOrFail($id);
+        $this->authorize('view', $user);
         $servers = $user->servers()
             ->paginate(config('constants.PAGINATION_LIST_LENGTH'));
 
@@ -60,6 +65,7 @@ class UserController extends Controller
      */
     public function edit(User $user)
     {
+        $this->authorize('update', $user);
         return view('users.edit', compact('user'));
     }
 
@@ -68,6 +74,7 @@ class UserController extends Controller
      */
     public function update(UpdateUserRequest $request, User $user)
     {
+        $this->authorize('update', $user);
 
         $userData = [
             'name' => $request->name,
@@ -75,17 +82,12 @@ class UserController extends Controller
             'role' => $request->role
         ];
 
-
         if($request->password)
         {
             $userData['password'] = Hash::make($request->password);
         }
 
-
-
         $user->update($userData);
-
-
         return redirect()
             ->route('users.show', $user->id)
             ->with('success', 'User updated successfully');
@@ -97,6 +99,7 @@ class UserController extends Controller
      */
     public function destroy(User $user)
     {
+        $this->authorize('delete', $user);
         $user->delete();
         return redirect()->route('users.index')->with('success', 'User deleted successfully');
     }
